chore: merge visual-report cli into test
